64bit: Fix lesser warnings, like float -> CGFloat, int -> NSInteger.

+++ Quicksilver/Code-App/QSFancyTableView.m
@@ -28,7 +28,7 @@
 
 - (void)drawBackgroundInClipRect:(NSRect)clipRect {
 	NSColor *backgroundColor = [self backgroundColor];
-	float hue, saturation, brightness, alpha;
+	CGFloat hue, saturation, brightness, alpha;
 	[[backgroundColor colorUsingColorSpaceName:NSDeviceRGBColorSpace] getHue:&hue saturation:&saturation brightness:&brightness alpha:&alpha];
 
 	// Create synthetic darker and lighter versions
@@ -58,7 +58,7 @@
 								// one row's selection can change how others draw.
 }
 
-- (void)deselectRow:(int)row; {
+- (void)deselectRow:(NSInteger)row; {
 	[super deselectRow:row];
 	[self setNeedsDisplay:YES];					   
 }
@@ -90,7 +90,7 @@
 
 - (void)drawBackgroundInClipRect:(NSRect)clipRect {
 	NSColor *backgroundColor = [self backgroundColor];
-	float hue, saturation, brightness, alpha;
+	CGFloat hue, saturation, brightness, alpha;
 	[[backgroundColor colorUsingColorSpaceName:NSDeviceRGBColorSpace] getHue:&hue saturation:&saturation brightness:&brightness alpha:&alpha];
 
 	// Create synthetic darker and lighter versions
@@ -124,7 +124,7 @@
 								//	one row's selection can change how others draw.
 }
 
-- (void)deselectRow:(int)row; {
+- (void)deselectRow:(NSInteger)row; {
 	[super deselectRow:row];
 	[self setNeedsDisplay:YES];						   
 }
